requirements.txt
# Probado con Python 3.11.9

numpy>=1.23,<2
pandas>=2.0,<3
rdkit-pypi
